Задание 5. Константные данные перенесены в отдельный модуль
