+++ js/data.js
@@ -1,0 +1,41 @@
+// Имена авторов комментариев
+const AUTHORS = [ 'Алиса',
+  'Варвара',
+  'Виктория',
+  'Адам',
+  'Иван',
+  'Кирилл',
+  'Роман',
+  'Мирослава',
+  'Максим',
+  'Ксения',
+  'Мария',
+  'Александр',
+  'Арина',
+  'Юлия',
+  'Данила',
+  'Михаил',
+  'Елизавета',
+  'Марк',
+  'Артём',
+  'Денис',
+  'Мирон',
+  'Пётр',
+  'Вероника',
+  'Константин',
+  'Алексей',
+  'Светлана',
+  'Ольга',
+  'Кира',
+  'Ярослав',
+  'Вячеслав'];
+
+// Массив предопределенных комментариев.
+const MESSAGES = [ 'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'];
+
+export {AUTHORS, MESSAGES};
